Rename transformer to transformerTemplate since it can accomodate LineStrings (or any valid GeoJSON geometry)
