+++ transformers/montana.js
@@ -1,5 +1,5 @@
 'use strict';
-var template = require('../lib/pointTransformer');
+var template = require('../lib/transformerTemplate');
 module.exports = template(function(props){
   return props.Number + ' ' + props.Address
 }, 'City', 'State', 'ZIP');
