+++ transformers/north_carolina.js
@@ -1,3 +1,3 @@
 'use strict';
-var template = require('../lib/pointTransformer');
+var template = require('../lib/transformerTemplate');
 module.exports = template('COMP_STR', '', 'STATE_NAME', 'ZIP_CODE');
